CERT-C_INT13: don't use bitwise ops with signed integers
More modifications for compliance.
Most of the mods regard flags management (e.g. save flags as unsigned
within the contexts)

+++ kernel/src/sys/sys_open.c
@@ -57,11 +57,11 @@ int sys_open(const char *pathname, int flags, mode_t mode)
     fil->ref = 1;
     fil->off = 0;
     fil->mode = mode;
-    fil->flags = flags & ~O_CLOEXEC;
+    fil->flags = (unsigned int)flags & ~O_CLOEXEC;
     fil->dent = dent;
 
     current->fds[fdn].fil = fil;
-    current->fds[fdn].flags = flags & O_CLOEXEC;
+    current->fds[fdn].flags = (unsigned int)flags & O_CLOEXEC;
 
     return fdn;
 }
